Created a separate package for fast external itemset
