+++ collect_app/src/main/java/org/odk/collect/android/provider/FormsProvider.java
@@ -27,7 +27,7 @@ import android.text.TextUtils;
 
 import androidx.annotation.NonNull;
 
-import org.odk.collect.android.database.ItemsetDbAdapter;
+import org.odk.collect.android.fastexternalitemset.ItemsetDbAdapter;
 import org.odk.collect.android.database.FormsDatabaseHelper;
 import org.odk.collect.android.provider.FormsProviderAPI.FormsColumns;
 import org.odk.collect.android.storage.StorageInitializer;
