+++ collect_app/src/main/java/org/odk/collect/android/storage/migration/StorageEraser.java
@@ -1,6 +1,6 @@
 package org.odk.collect.android.storage.migration;
 
-import org.odk.collect.android.database.ItemsetDbAdapter;
+import org.odk.collect.android.fastexternalitemset.ItemsetDbAdapter;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
