+++ collect_app/src/main/java/org/odk/collect/android/utilities/ApplicationResetter.java
@@ -20,7 +20,7 @@ import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.configure.ServerRepository;
 import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.dao.InstancesDao;
-import org.odk.collect.android.database.ItemsetDbAdapter;
+import org.odk.collect.android.fastexternalitemset.ItemsetDbAdapter;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.logic.PropertyManager;
 import org.odk.collect.android.preferences.AdminSharedPreferences;
